% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calibrate_channel}
\alias{calibrate_channel}
\title{Calibrate channel data}
\usage{
calibrate_channel(channel.name, data, channel.list)
}
\arguments{
\item{channel.name}{The name of the channel}

\item{data}{The channel data as a vector}

\item{channel.list}{A list containing channel information (from the function \code{get_channels}).}
}
\value{
The calibrated channel data.
}
\description{
Calibrate channel data
}
\keyword{internal}

