% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{channel_name_number}
\alias{channel_name_number}
\title{Translate between channel name and number.}
\usage{
channel_name_number(channel, channel.list)
}
\arguments{
\item{channel}{The number or name of a channel.}

\item{channel.list}{A list containing channel information (from the function \code{get_channels}).}
}
\value{
A list with the name and number of the channel.
}
\description{
Given a channel name or a channel number, find the corresponding number or name.
}
\seealso{
get_channels
}
\keyword{internal}

