% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{data_matrix_to_list}
\alias{data_matrix_to_list}
\title{Convert a data matrix to a list}
\usage{
data_matrix_to_list(data, channel_list, samplingrate)
}
\arguments{
\item{data}{A matrix containing the channel data in the columns}

\item{data}{}
}
\value{
A list with the data channels.
}
\description{
Convert a data matrix to a list
}
\keyword{internal}

