% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_channels}
\alias{get_channels}
\title{Read channel information}
\usage{
get_channels(protocol)
}
\arguments{
\item{protocol}{The protocol as a list.}
}
\value{
A list with with the channel information (only TableProtocolInputs are kept in the input list).
}
\description{
Read information on the channels.
}
\keyword{internal}

