% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_data_files}
\alias{get_data_files}
\title{Return a list of data files related to the given session number.}
\usage{
get_data_files(datadir, session.number = 1)
}
\arguments{
\item{datadir}{The directory containing the neurOne data}

\item{session.number}{The session number. Default is 1.}
}
\value{
The data files as a list.
}
\description{
Return a list of data files related to the given session number.
}
\keyword{internal}

