% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_event_structure}
\alias{get_event_structure}
\title{Define the structure of a neurOne event.}
\usage{
get_event_structure()
}
\value{
A list describing the structure of the event.
}
\description{
Define the structure of a neurOne event.
}
\keyword{internal}

