% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_number_of_sessions}
\alias{get_number_of_sessions}
\title{Return the number of sessions in the recording}
\usage{
get_number_of_sessions(session)
}
\arguments{
\item{session}{The session as a list.}
}
\value{
The number of sessions in the recordiing.
}
\description{
Return the number of sessions in the recording
}
\keyword{internal}

