% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_protocol}
\alias{get_protocol}
\title{Read protocol information.}
\usage{
get_protocol(protocol)
}
\arguments{
\item{protocol}{The protocol as a list.}
}
\value{
A list with with the protocol (inputs, vurves and monitors are removed from the input list).
}
\description{
Read the protocol information from a list structure containing the protocol
}
\keyword{internal}

