% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_session_start_and_end}
\alias{get_session_start_and_end}
\title{Get the start and end times of a particular session.}
\usage{
get_session_start_and_end(session, session_number = 1)
}
\arguments{
\item{session}{A list with the session information (read from an XML file).}

\item{session_number}{The session number. Default is 1.}
}
\value{
The start and stop times (raw and POSIXct) of the session.
}
\description{
Get the start and end times of a particular session.
}
\keyword{internal}

