% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neurone.R
\docType{package}
\name{neurone}
\alias{neurone}
\alias{neurone-package}
\title{Read Physiologic Signal Data Recorded with a Mega Electronics neurOne Device.}
\description{
The neurone package contains functions for reading data from a Mega
Electronics Ltd neurOne device.
}

