% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_event}
\alias{parse_event}
\title{Parse a neurOne event.}
\usage{
parse_event(data, event_structure)
}
\arguments{
\item{data}{A list of 88 bytes containing the event data.}

\item{eventd_.structure}{The structure of the event. See \code{\link{get_event_structure}}}
}
\value{
The source port as a string.
}
\description{
Parse a neurOne event.
}
\keyword{internal}

