% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_timestamp}
\alias{parse_timestamp}
\title{Convert a string to a timestamp.}
\usage{
parse_timestamp(ts, format = "\%Y-\%m-\%dT\%H:\%M:\%S")
}
\value{
A timestamp representing the given string.
}
\description{
Convert a string to a timestamp.
}

