% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neurone.R
\name{read.neurone}
\alias{read.neurone}
\title{Read data from a Mega Electronics neurOne device.}
\usage{
read.neurone(datadir, session.number = 1, start = 0, data.length = NULL,
  channels = NULL, read.events = TRUE, header.only = FALSE)
}
\arguments{
\item{datadir}{The directory containing the data.}

\item{session.number}{The session number. Default is 1.}

\item{start}{The offset in seconds from the beginning of the file to start reading data.}

\item{data.length}{The amount of data in seconds to be read.}

\item{channels}{A list containing channel names to be read.}

\item{read.events}{Should events be read. Boolean. Default is TRUE.}

\item{header.only}{Should only the header information, but no signal data, be read. Boolean. Default is FALSE.}
}
\value{
A recording structure (a list) containing the data.
}
\description{
Read data from a Mega Electronics neurOne device.
}

