% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_channel}
\alias{read_channel}
\title{Read binary neurOne data from one channel.}
\usage{
read_channel(binfile, channel, channel.list, samplingrate, start = 0,
  data.length = NULL)
}
\arguments{
\item{binfile}{The name of the file containing the data.}

\item{channel}{The number or name of the channel to be read.}

\item{channel.list}{A list containing channel information (from the function \code{get_channels}).}

\item{samplingrate}{The sampling rate.}

\item{start}{The offset in seconds from the beginning of the file to start reading data.}

\item{data.length}{The amount of data in seconds to be read.}
}
\value{
A matrix with the channel data.
}
\description{
Read binary neurOne data from a bin-file. Only data from one channel is read.
}
\seealso{
get_channels
}
\keyword{internal}

