% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_channels}
\alias{read_channels}
\title{Read binary neurOne data from multiple channels.}
\usage{
read_channels(binfile, channels, channel.list, samplingrate, start = 0,
  data.length = NULL)
}
\arguments{
\item{binfile}{The name of the file containing the data.}

\item{channels}{The number or name of the channels to be read.}

\item{channel.list}{A list containing channel information (from the function \code{get_channels}).}

\item{samplingrate}{The sampling rate.}

\item{start}{The offset in seconds from the beginning of the file to start reading data.}

\item{data.length}{The amount of data in seconds to be read.}
}
\value{
A matrix with the channel data.
}
\description{
Read binary neurOne data from a bin-file. Data from multiple channels is read.
}
\seealso{
get_channels
}
\keyword{internal}

