% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_events}
\alias{read_events}
\title{Read events from a binary neurOne event file.}
\usage{
read_events(datadir, session.number = 1, samplingrate = NULL)
}
\arguments{
\item{datadir}{The directory containing the neurOne data.}

\item{session.number}{The session number. Default is 1.}

\item{samplingrate}{The sampling rate.}
}
\value{
A matrix with the channel data.
}
\description{
Read events from a binary neurOne event file.
}
\seealso{
get_channels
}
\keyword{internal}

