% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_session_data}
\alias{read_session_data}
\title{Read all of the channel data associated with a particular session.}
\usage{
read_session_data(datadir, channel.list, channels, samplingrate, start,
  data.length, session.number = 1)
}
\arguments{
\item{datadir}{The directory containing the neurOne data.}

\item{channel.list}{A list containing channel information (from the function \code{get_channels}).}

\item{samplingrate}{The sampling rate.}

\item{start}{The offset in seconds from the beginning of the file to start reading data.}

\item{data.length}{The amount of data in seconds to be read.}

\item{session.number}{The session number. Default is 1.}
}
\value{
A matrix with the channel data.
}
\description{
Read all of the channel data associated with a particular session.
}
\seealso{
get_channels
}
\keyword{internal}

