% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_code_event_type}
\alias{set_code_event_type}
\title{Map the numeric event type to a string and set the event code.}
\usage{
set_code_event_type(event, files.events)
}
\arguments{
\item{event}{An event (as a one-row data frame)}

\item{files.events}{A list containing the event files. See \code{\link{get_event_files}}}
}
\value{
The event with the type and code modified.
}
\description{
Also read user-defined events from the file eventData.bin and set
the contents as the event type.
}
\keyword{internal}

