% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_source_port}
\alias{set_source_port}
\title{Map a numeric source port to a string representation.}
\usage{
set_source_port(x)
}
\arguments{
\item{x}{A numeric describing the source port.}
}
\value{
The source port as a string.
}
\description{
Map a numeric source port to a string representation.
}
\keyword{internal}

