% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{xml_to_list}
\alias{xml_to_list}
\title{#' Read an XML file and retun the result as a list.}
\usage{
xml_to_list(filename)
}
\arguments{
\item{filename}{The path to the XML file.}
}
\value{
The XML file as a list.
}
\description{
Read an XML file and retun the result as a list.
}
\keyword{internal}

